+++ inc/Core/Conditions/Condition.h
@@ -1,7 +1,7 @@
 #ifndef CONDITION_H
 #define CONDITION_H
 
-#include "../Config/Logger.h"
+#include "../../Config/Logger.h"
 
 #include <string>
