+++ inc/Core/Responses/Response.h
@@ -1,0 +1,9 @@
+#ifndef RESPONSE_H 
+#define RESPONSE_H
+
+// INFO: this class is all about the response classes from different classes across the project.
+// meaning -> from db model a model returns a certain reponse from the db, due to that we will
+// create a class that will handle different reponse cases and return the right response, i.e:
+// virtual DBResponse WriteToDB(); -> it will hold data, codes, general info about the response.
+
+#endif
