+++ tests/Test.h
@@ -1,0 +1,6 @@
+#ifndef TEST_H
+#define TEST_H
+
+#include <catch.hpp>
+
+#endif
